requirements.txt
nextcord
python-dotenv
colorama
